requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dj-database-url==1.0.0
Django==4.1.1
django-bootstrap-form==3.4
django-environ==0.9.0
django-filter==22.1
django-on-heroku==1.1.2
exceptiongroup==1.0.4
gunicorn==20.1.0
h11==0.14.0
idna==3.4
iniconfig==1.1.1
npm==0.1.1
optional-django==0.1.0
outcome==1.2.0
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.5
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-environ==0.4.54
requests==2.28.1
selenium==4.6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
vercel==0.2.1
whitenoise==6.2.0
wsproto==1.2.0
